Remove additional rule registration
